Pass test

+++ RLETest.cpp
@@ -6,6 +6,6 @@ using namespace std;
 
 TEST(RLETest, BasicAssertions)
 {
-    vector<pair<char, int>> expected_visited = {{'a', 1}, {'b', 2}, {'c', 3}};
+    vector<pair<char, long long>> expected_visited = {{'a', 1}, {'b', 2}, {'c', 3}};
     EXPECT_EQ(expected_visited, rle("abbccc"));
 }
